Develop Verify Annotations .py
I'm getting this preliminary analysis ready so that I can verify that the standard deviation of the data can be used to verify that the annotation markers are properly set. Stopped at the point of creating data segments and calculating standard deviation within those segments. Moving forward I need to alter the protocol.py script to collect data in a way that allows us to check for activity (touching a given channel and not others).