Added Save Functionality to the End_Connection Function in the Connect Object
The end_connection method now can be run with save true or save false. If save is true it will be important to provide a save_as (filename -- no extension included). The fif.gz extension is default. I also need to update the channel locations here, because I just used some old ones for a proof of concept.