Print Data Values in the best location to record them
I went farther and even returned an epoch_info dictionary from the present_slide_set() method. This is only meant for slide_sets which deal with condition related epochs, but the only condition which is equipped to create the epoch_info object at this point is SSVEP. To equip the other slide sets in the future, add code to create a start_time and duration variable which will then populate the epoch_info object.