Calculate PSD for alpha wave check
Calculate the power-spectral density for a period of time while the user has their eyes closed and compare it to  a period of time when the user doesn't have their eyes closed. Ask the proctor (with a dialog box) if the ratio of alpha wave presence to not-present is reasonable.